// SPM2021-MWE-sprint3-front-end (1)/SPM2021-MWE-sprint3-front-end/SPM2021/front-end/src/app/model/geoPoint.ts
import { IotLiteCoverage } from "./iotLiteCoverage";
import { IotLiteObject } from "./iotLiteObject";
import { SsnDeployment } from "./ssnDeployment";

export class GeoPoint {
    id: number;
    iotliteAltRelative: string;    
    longg: string;    
    iotliteObject: IotLiteObject;    
    ssnDeployment: SsnDeployment;    
    iotlitecoverage: IotLiteCoverage;    
    iotlitealt: String;    
    iotliteRelativeLocation: String;    

}